refactor: 예약 DB 접근 로직 Repository으로 이동
