+++ src/main/java/roomescape/repository/entity/ReservationTimeEntity.java
@@ -1,0 +1,9 @@
+package roomescape.repository.entity;
+
+import java.time.LocalTime;
+
+public record ReservationTimeEntity(
+    Long id,
+    LocalTime startAt
+) {
+}
