+++ src/main/java/roomescape/repository/entity/ReservationWithTimeEntity.java
@@ -1,0 +1,11 @@
+package roomescape.repository.entity;
+
+import java.time.LocalDate;
+
+public record ReservationWithTimeEntity(
+        Long id,
+        String name,
+        LocalDate date,
+        ReservationTimeEntity time
+) {
+}
